NotesApp/requirements.txt
openai
django
djangorestframework